backend/requirements.txt
python-dotenv
numpy
pandas
tensorflow
pillow
scikit-learn
google-generativeai
pdfminer.six
torch
torchvision
fastapi
uvicorn
python-multipart
requests
